Migrating plugin to new format
